+++ epyk/core/css/styles/classes/CssStylesDivEvents.py
@@ -36,8 +36,8 @@ class CssDivOnHoverBackgroundLight(CssStyle.Style):
       "background-color": self.page.theme.colors[0],
       'border': "1px solid %s" % self.page.theme.greys[0]})
     self.hover.css({
-      'border': "1px solid %s" % self.page.theme.notch(),
-      "background-color": self.page.theme.notch(), "text-decoration": 'underline'})
+      'border': "1px solid %s" % self.page.theme.notch(-4),
+      "background-color": self.page.theme.notch(-4), "text-decoration": 'underline'})
     self.hover.css('color', self.page.theme.greys[0], important=True)